+++ commands/run.go
@@ -18,7 +18,6 @@ import (
 )
 
 type multiFlags []string
-type externalCommand []string
 
 func (sf *multiFlags) String() string {
 	return strings.Join(*sf, ", ")
